+++ src/condor_shadow.V6.1/remoteresource.cpp
@@ -1603,9 +1603,9 @@ RemoteResource::recordCheckpointEvent( ClassAd* update_ad )
 		job_committed_time += now - int_value;
 		jobAd->Assign(ATTR_JOB_COMMITTED_TIME, job_committed_time);
 
-		float slot_weight = 1;
+		double slot_weight = 1;
 		jobAd->LookupFloat(ATTR_JOB_MACHINE_ATTR_SLOT_WEIGHT0, slot_weight);
-		float slot_time = 0;
+		double slot_time = 0;
 		jobAd->LookupFloat(ATTR_COMMITTED_SLOT_TIME, slot_time);
 		slot_time += slot_weight * (now - int_value);
 		jobAd->Assign(ATTR_COMMITTED_SLOT_TIME, slot_time);
